move: unassigned card componet shared 폴더로 이동
